New MismatchError type for auditing, collection logic updated
Mismatch Error:
Created a new Exception type called "MismatchError" that should always be checked prior to returning a new df, and particularly so if about to overwrite a data file. It checks expected values against actual values (making sure there are the correct number of observations) 

Collection Logic Updated:
Check if an episodes_core-YYYY-MM-DD.csv file exists. If it doesn't, collect everything from scratch. If it does exist, check 'YYYY-MM-DD' to see if it was collected fewer than 20 days ago. If it was, re-collect only the 5 most recent episodes (and drop duplicates). Otherwise, assume that this is very out of date and recollect everything.