getEpCompletionRate()
This function collects the average completion rate of listens on the *SIMPLECAST WEB PLAYER* (ie. this is not necessarily a representative sample). It only computes the average completion rate at a given time so we never throw away old data. We always append new data so that we can track how this number changes over time for each episode.