Fixing fileCleanup()
In `apiCollect.py` we developed a problem in failing to remove old files from the DropBox folder. This update fixes that. 

The failure was caused by having an invalid pattern that we were trying to match. This should let fileCleanup() run all the way through and clean up the DropBox after each daily collection.