Updating the Docker Structure for deployment
We are revising the structure of the Docker deployment to make debugging/updating faster (conditional on not using new R packages).

Formerly a single docker image would be built to re-create the entire R environment (with packages, using .renv) and then running the deployment script (`deployApp.R`). However, in Linux, renv requires the recompiling of every package at build time, making the build of this Docker image painfully slow. Making any change in the environment/deployment script would require a near 20-minute build wait.

This commit makes two big changes. First it splits the R environment (package) construction into its own dockerfile called `localRenvRestoreDockerfile.dockerfile`. This does a lot of the heavy lifting of the app because it uses to renv.lock file to reconstruct the R environment as appropriate. This only needs to be done once. This image is saved as 'renv-restore'

In the second part, a second dockerfile with the default name ("Dockerfile") simply runs the deployment script after loading the 'renv-restore' image (with all of the already-compiled projects).

NB: If the shiny app/dashboard needs to use any new packages that aren't already in the .renv.lock file, the first Docker image needs to be re-built. Otherwise, only the second image needs to be run.

TODO: Add the above to the README.md noting the different GitHub/local deployment workflows.