Function for better local/remote development and deployment
In the past, I just relied on commenting out certain blocks of code to prepare the app files for local development vs preparing it for deployment.

Now there is a function that is just called at the top of global.R that switches up the way data files are accessed for local development vs remote deployment.