Updating automateCollect.bat
Windows Task Schedule executes this batch file. Inside this file, new code has been written that has all of the output of apiCollect.py (whatever is usually being printed to the console) printed to a .txt file named with the date so that we can look up specific date failures.