More verbose logging and a bug fix
The simplecast API recently changed the way it is returning some URLs causing a df length mismatch error. I have switched to dropping duplicates on `episode_id` instead of the URL to hopefully head this off in the future.

Also more verbose printing for better logging and bug squashing in the future